+++ addons/fray/src/input/device/input_data/binds/input_bind_action.gd
@@ -13,7 +13,7 @@ func _init(action_name: String = "") -> void:
 	action = action_name
 
 
-func _is_pressed_impl(device: int = 0) -> bool:
+func _is_pressed_impl(_device: int = 0) -> bool:
 	if not InputMap.has_action(action):
 		push_warning("Action '%s' does not exist in Godot InputMap" % action)
 		return false
@@ -24,3 +24,7 @@ func _equals_impl(input_bind: Resource) -> bool:
 	return (
 		._equals_impl(input_bind)
 		and action == input_bind.action)
+
+
+func _get_strength_impl(_device: int = 0) -> float:
+	return Input.get_action_strength(action)