+++ addons/fray/src/input/device/input_data/binds/input_bind_key.gd
@@ -9,7 +9,7 @@ func _init(keyboard_key: int = -1) -> void:
 	key = keyboard_key
 
 	
-func _is_pressed_impl(device: int = 0) -> bool:
+func _is_pressed_impl(_device: int = 0) -> bool:
 	return Input.is_key_pressed(key)