+++ addons/fray/src/input/device/input_data/binds/input_bind_mouse_button.gd
@@ -9,7 +9,7 @@ func _init(mouse_button: int = -1) -> void:
 	button = mouse_button
 
 
-func _is_pressed_impl(device: int = 0) -> bool:
+func _is_pressed_impl(_device: int = 0) -> bool:
 	return Input.is_mouse_button_pressed(button)