+++ addons/fray/src/input/device/input_data/state/input_state.gd
@@ -10,6 +10,7 @@ var virtually_pressed: bool
 var physics_frame: int = -1
 var idle_frame: int = -1
 var time_pressed: int = -1
+var strength: float
 
 
 func press(is_virtual_press: bool = false) -> void:
@@ -19,8 +20,12 @@ func press(is_virtual_press: bool = false) -> void:
     time_pressed = OS.get_ticks_msec()
     virtually_pressed = is_virtual_press
 
+    if strength <= 0:
+        strength = 1
+
 
 func unpress() -> void:
     pressed = false
     physics_frame = Engine.get_physics_frames()
     idle_frame = Engine.get_idle_frames()
+    strength = 0